Pull proportion morphometrically mature Chionoecetes spp. males by size

A time series of model estimated proportion of morphometrically mature male Tanner Crab and Snow Crab by 10 millimeter size bin for a given region or district, based on chela height measurements. Yearly model parameter estimates of 50% probability of maturity at size are also provided.

Usage

get_male_maturity(
  species = NULL,
  region = c("EBS", "NBS")[1],
  district = NULL,
  channel = NULL
)

Arguments

species

Character string. One of c("TANNER", "SNOW").

region

Character string describing the region of interest. One of c("EBS", "NBS"). Defaults to "EBS" for Eastern Bering Sea.

district

Character string. One or many of c("ALL", "E166", "W166"). Defaults to "ALL"; "E166" and "W166" are used for Tanner Crab only.

channel

Connection to Oracle created via crabpack::get_connected() or RODBC::odbcConnect().

Value

A named list containing the proportion of male Chionoecetes spp. crab that are morphometrically mature in a given 10mm size bin, and yearly model parameter estimates for 50% probability of maturity at size for the species, region, and district of interest.